JAMES-2525 move swift specific types to swift subpackage
