+++ server/blob/blob-objectstorage/src/main/java/org/apache/james/blob/objectstorage/swift/UserHeaderName.java
@@ -17,47 +17,14 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.blob.objectstorage;
+package org.apache.james.blob.objectstorage.swift;
 
-import com.google.common.base.MoreObjects;
-import com.google.common.base.Objects;
-
-public final class TenantName {
-    public static TenantName of(String value) {
-        return new TenantName(value);
+public final class UserHeaderName extends HeaderName {
+    public static UserHeaderName of(String value) {
+        return new UserHeaderName(value);
     }
 
-    private final String value;
-
-    private TenantName(String value) {
-        this.value = value;
-    }
-
-    public String value() {
-        return value;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        TenantName that = (TenantName) o;
-        return Objects.equal(value, that.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(value);
-    }
-
-    @Override
-    public String toString() {
-        return MoreObjects.toStringHelper(this)
-            .add("value", value)
-            .toString();
+    private UserHeaderName(String value) {
+        super(value);
     }
 }
