+++ server/blob/blob-objectstorage/src/test/java/org/apache/james/blob/objectstorage/swift/CredentialsTest.java
@@ -17,15 +17,15 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.blob.objectstorage;
+package org.apache.james.blob.objectstorage.swift;
 
 import org.junit.jupiter.api.Test;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
 
-class RegionTest {
+class CredentialsTest {
     @Test
-    public void regionShouldRespectBeanContract() {
-        EqualsVerifier.forClass(Region.class).verify();
+    public void credentialsShouldRespectBeanContract() {
+        EqualsVerifier.forClass(Credentials.class).verify();
     }
 }
